+++ src/script/bindings/mod.rs
@@ -8,3 +8,4 @@ pub mod ping;
 pub mod redis;
 pub mod tcp;
 pub mod tls;
+pub mod ws;